feat: refactor trustfall_wasm test suite
