+++ .idea/trustfall.iml
@@ -1,0 +1,25 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module type="EMPTY_MODULE" version="4">
+  <component name="NewModuleRootManager">
+    <content url="file://$MODULE_DIR$">
+      <sourceFolder url="file://$MODULE_DIR$/demo-hytradboi/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/experiments/schemaless/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/experiments/schemaless_wasm/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/pytrustfall/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall/examples" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall/tests" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall_core/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall_derive/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall_derive/tests" isTestSource="true" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall_filetests_macros/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall_stubgen/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall_testbin/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall_wasm/src" isTestSource="false" />
+      <sourceFolder url="file://$MODULE_DIR$/trustfall_wasm/tests" isTestSource="true" />
+      <excludeFolder url="file://$MODULE_DIR$/target" />
+    </content>
+    <orderEntry type="inheritedJdk" />
+    <orderEntry type="sourceFolder" forTests="false" />
+  </component>
+</module>
